+++ database/migrations/2021_02_20_042943_create_kasirs_table.php
@@ -18,7 +18,7 @@ class CreateKasirsTable extends Migration
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
             $table->char('kode_user', 4)->unique();
             $table->integer('number_of_transaction')->default(0);
-            $table->bigInteger('total_penjualan')->nullable();
+            $table->double('total_penjualan')->nullable();
             $table->timestamps();
         });
     }
