+++ database/migrations/2021_02_24_011441_create_keuangan_businesses_table.php
@@ -16,9 +16,9 @@ class CreateKeuanganBusinessesTable extends Migration
         Schema::create('keuangan_businesses', function (Blueprint $table) {
             $table->id();
             $table->foreignId('business_id')->constrained('businesses');
-            $table->bigInteger('pemasukan')->default(0);
-            $table->bigInteger('pengeluaran')->default(0);
-            $table->bigInteger('saldo')->default(10000000);
+            $table->double('pemasukan')->default(0);
+            $table->double('pengeluaran')->default(0);
+            $table->double('saldo')->default(10000000);
             $table->timestamps();
         });
     }
